fix(toast): Move ToastItemContext to root toast directory
