+++ src/toast/close-button/ToastCloseButton.tsx
@@ -3,7 +3,7 @@ import "./_toast-close-button.scss";
 import React from "react";
 
 import Button from "../../button/Button";
-import {useToastCloseButton, useToaster} from "../util/toastHooks";
+import {useToastItemContext, useToaster} from "../util/toastHooks";
 
 interface ToastCloseButtonProps {
   children: React.ReactNode;
@@ -12,7 +12,7 @@ interface ToastCloseButtonProps {
 
 function ToastCloseButton({children, testid}: ToastCloseButtonProps) {
   const {hide} = useToaster();
-  const {toastId} = useToastCloseButton();
+  const {toastId} = useToastItemContext();
 
   return (
     <Button testid={testid} onClick={handleClick} customClassName={"toast-close-button"}>
